backend/requirements.txt
# Core Framework
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
python-multipart

# Database & Storage
supabase
httpx

# Auth
python-jose[cryptography]
passlib[bcrypt]

# AI Providers
google-generativeai
fal-client

# Upstash Redis (Serverless-compatible)
upstash-redis

# Monitoring & Logging
sentry-sdk

# Testing (dev only)
pytest
